requirements.txt
flask
pytesseract
gtts
opencv-python
Pillow
numpy
